fix: more specta fixes

+++ src/traits/discovery.rs
@@ -32,10 +32,10 @@ pub struct DiscoveryQuery {
 #[derive(Debug, Clone, Serialize, Deserialize)]
 #[cfg_attr(feature = "specta", derive(specta::Type))]
 pub struct PaginationMeta {
-    pub current: u64,
-    pub page_size: u64,
-    pub total_pages: Option<u64>,
-    pub total_items: Option<u64>,
+    pub current: u32,
+    pub page_size: u32,
+    pub total_pages: Option<u32>,
+    pub total_items: Option<u32>,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
@@ -69,9 +69,9 @@ pub struct ModSummary {
     pub name: String,
     pub description: String,
     pub short_description: String,
-    pub downloads: u64,
-    pub views: u64,
-    pub likes: u64,
+    pub downloads: u32,
+    pub views: u32,
+    pub likes: u32,
     pub thumbnail_image: String,
     pub tags: Vec<String>,
     pub user_name: String,
